[feature/#9] write activity 뷰 수정
