--- app/src/main/java/org/sopt/seminar/writeData.kt
@@ -1,8 +1,0 @@
-package org.sopt.seminar
-
-data class writeData (
-    val title: String,
-    val category: String,
-    val price: Int,
-    val content: String
-)
